Rename and move PayPalVaultEditPendingRequest to vaultedit directory
