+++ PayPal/src/main/java/com/braintreepayments/api/paypal/vaultedit/PayPalVaultEditPendingRequest.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.paypal
+package com.braintreepayments.api.paypal.vaultedit
 
 /**
  * A pending request for the PayPal web-based authentication flow created by invoking
@@ -6,7 +6,7 @@ package com.braintreepayments.api.paypal
  * on-device and used to deliver a result of the browser flow in
  * [PayPalLauncher.handleReturnToApp]
  */
-sealed class PayPalEditPendingRequest {
+sealed class PayPalVaultEditPendingRequest {
 
     /**
      * A pending request was successfully started.
@@ -14,10 +14,10 @@ sealed class PayPalEditPendingRequest {
      * @property pendingRequestString - This String should be stored and passed to
      * [PayPalLauncher.handleReturnToApp].
      */
-    class Started internal constructor(val pendingRequestString: String) : PayPalEditPendingRequest()
+    class Started internal constructor(val pendingRequestString: String) : PayPalVaultEditPendingRequest()
 
     /**
      * An error occurred launching the PayPal browser flow. See [error] for details.
      */
-    class Failure internal constructor(val error: Exception) : PayPalEditPendingRequest()
+    class Failure internal constructor(val error: Exception) : PayPalVaultEditPendingRequest()
 }
